requirements.txt
-e git+https://github.com/JoseChipanaTica/ai-website-generator.git@3f6261ef8b7667fdf45e2114d9a4c7e04b901763#egg=ai_website_generator
ai21==1.0.5
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
fastapi==0.95.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
langchain==0.0.147
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openapi-schema-pydantic==1.2.4
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
uuid==1.30
uvicorn==0.21.1
yarl==1.9.1
redis~=4.5.4
